requirements.txt
appdirs==1.4.4
certifi==2021.10.8
charset-normalizer==2.0.8
distlib==0.3.1
filelock==3.0.12
idna==3.3
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
urllib3==1.26.7
virtualenv==20.4.2
